requirements.txt
fastapi
uvicorn
pydantic
kafka-python
langchain
langgraph
dspy-ai
boto3
psycopg2-binary
sqlalchemy
python-dotenv
alembic